fps float

+++ tracker/animal/tracker.py
@@ -168,7 +168,7 @@ class AnimalTrackerKalman(AnimalTracker_CPU):
 
     def __init__(
             self, 
-            fps: int, 
+            fps: float, 
             model_order: int, 
             model_uncertainty: float = 0.2,
             measurement_uncertainty: float = 1.0,
